refactor: remove arch package
