+++ app/src/main/java/com/monstarlab/core/extensions/ViewErrorAware.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.arch.extensions
+package com.monstarlab.core.extensions
 
 interface ViewErrorAware
