+++ app/src/main/java/com/monstarlab/core/extensions/ViewExtensions.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.arch.extensions
+package com.monstarlab.core.extensions
 
 import android.view.View
 import androidx.core.view.isVisible
@@ -8,6 +8,7 @@ import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import com.google.android.material.snackbar.Snackbar
 import com.monstarlab.core.sharedui.errorhandling.ViewError
+import com.monstarlab.core.utils.DebounceOnClickListener
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.*
