+++ app/src/main/java/com/monstarlab/core/injection/modules/OkHttpModule.kt
@@ -17,7 +17,7 @@ import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module
-class RestModule {
+class OkHttpModule {
 
     @Provides
     @Singleton
