+++ app/src/main/java/com/monstarlab/core/persistence/SharedPreferenceDataStore.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.arch.data
+package com.monstarlab.core.persistence
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
