+++ app/src/main/java/com/monstarlab/core/utils/DebounceOnClickListener.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.arch.extensions
+package com.monstarlab.core.utils
 
 import android.view.View
