+++ app/src/main/java/com/monstarlab/core/utils/GlobalViewBindingDelegate.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.arch.extensions
+package com.monstarlab.core.utils
 
 import android.view.View
 import androidx.viewbinding.ViewBinding
