+++ app/src/main/java/com/monstarlab/features/resources/domain/ResourceRepository.kt
@@ -1,7 +1,7 @@
 package com.monstarlab.features.resources.domain
 
-import com.monstarlab.arch.data.Repository
-import com.monstarlab.arch.extensions.mapSuccess
+import com.monstarlab.core.domain.Repository
+import com.monstarlab.core.extensions.mapSuccess
 import com.monstarlab.features.resources.data.toEntity
 import com.monstarlab.features.resources.data.ResourcePreferenceStore
 import com.monstarlab.features.resources.data.ResourcesApi
