+++ app/src/main/java/com/monstarlab/features/user/data/UserPreferenceStore.kt
@@ -2,7 +2,7 @@ package com.monstarlab.features.user.data
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
-import com.monstarlab.arch.data.SingleSharedPreferenceDataStore
+import com.monstarlab.core.persistence.SingleSharedPreferenceDataStore
 import com.monstarlab.features.user.domain.User
 import javax.inject.Inject
